fixing a few gcc warnings
Former-commit-id: 3e31ae4be4b924394f913315fc07e1d43db982bd

+++ applications/plugins/Flexible/material/HookeForceField.h
@@ -101,7 +101,7 @@ public:
          typedef component::engine::BaseGaussPointSampler::volumeIntegralType volumeIntegralType;
          vector<volumeIntegralType> _volIntType;
          _volIntType.resize(_vol.size());
-         for(int i=0; i<_volIntType.size(); ++i)
+         for(size_t i=0; i<_volIntType.size(); ++i)
          {
              _volIntType[i].resize(1);
              _volIntType[i][0] = _vol[i];
